feat: 1543 refactor route structure and fixes for CR
